Change tests/api/dimensions_test.py to module scope.
